was supposed to make python scripts not files using __main__ == __name__

+++ 0x02-python-import_modules/calculation_1.py
@@ -14,4 +14,4 @@ def mul(a, b):
 
 # divides input a b
 def div(a, b):
-    return (a / b)
+    return (int(a / b))
